@@ -1,25 +1,5 @@
-import Vue from 'vue'
+import {createApp} from 'vue'
 import App from './App.vue'
-import Cadastro from './Cadastro.vue'
+import router from './router'
 
-const NotFound = { template: '<p>Página não encontrada</p>'}
-const Home = App;
-const CadastroPage = Cadastro;
-
-const routes = {
-  '/': Home,
-  '/cadastro': CadastroPage
-}
-
-new Vue({
-  el: '#app',
-  data: {
-    currentRoute: window.location.pathname
-  },
-  computed: {
-    ViewComponent () {
-      return routes[this.currentRoute] || NotFound
-    }
-  },
-  render (h) { return h(this.ViewComponent) }
-})
+createApp(App).use(router).mount('#app')
